Added options to enable auto on/off based on ambient lux and time
By setting this to non-zero on timeout, the display will turn off
if the ambient light is below a certain level for a customizable amount
of time, it will turn off the display.

Once it's off, it works in reverse. Ambient light must be above for
the same amount of time to turn on.

The auto on/off will still be bound by the schedule, so no matter how
much light there is during OFF hours, the display will remain off.

NOTE! Actual feature is yet to be implemented, but this adds the necessary
settings for this to be controlled.