Fixed issue where gif animations would cause out-of-memory scenarios
Google will sometimes produce 20MB+ gif animations out of your photos
if you, for example, upload a long sequence of similar images. This
caused convert to try and operate on each frame, eating up a massive
amount of memory.

Now we just grab the FIRST frame from a gif animation and process that.