Internal display support
* Issues with power management

- Was unable to change off time due to copy'n'paste
- Was unable to turn off scheduling

* Initial version of RGB888<->RGB565 conversion tool

* Initial change to support non-HDMI displays

* Reframing limits

Reframing should not happen if zoomed area is smaller than 20px on each
side of the frame. Looks silly.

* Major rewrite to tvservice handling

No longer deals with pixels, uses tv service string to deduce all
values, avoids trying to use non-existant resolution if TV is unplugged.

* Handle lack of ANY display

photoframe should not crash if a display is missing, instead UX should
show the issue at hand.

* Stop exception when no colorsensor is available

* Handle broken json from google

Sometimes downloads fail and the JSON is broken, which we now deal
with, but the rest of the code doesn't. This fixes the issue.

* Initial support for uploading drivers

Also enables listing and activating them

* Added support for uploading drivers

* Better handling of no display at all

* Reboot/power off now works as expcted

Turns out that replacing the HTML during async call will cancel
out that call. Now we just replace the body instead of the document.

* Add driver documentation

* Removed usage of old setting name

resolution is gone, replaced with tvservice

* Relocate all config files into separate folder

* Corrected README.md since it used old paths

Also clarified the need for chmod