Update MIGRATION.md
I took the 2019-01-05 image, freshly installed and upgraded to a clean current version on RPi3B.  Then tried the instructions.   pip3 needed to be installed, and in this distro, python3-flask-httpauth was not a separate package.   I thought it might be missing and tried installing with pip3, but it claimed that it was already there from the dist-packages.    BTW, do you have a flag for the update script to select a branch, or just use git --branch ?